refactor(rating-point): widen value type from byte to int

+++ backend/services/rating-service/rating/src/main/java/vn/uit/edu/msshop/rating/adapter/in/web/request/PostRatingRequest.java
@@ -8,5 +8,5 @@ public record PostRatingRequest(
         UUID userId,
         String userName,
         String userAvatar,
-        byte ratingPoint) {
+        int ratingPoint) {
 }
