+++ backend/services/rating-service/rating/src/main/java/vn/uit/edu/msshop/rating/adapter/in/web/response/RatingResponse.java
@@ -12,5 +12,5 @@ public record RatingResponse(
         UUID userId,
         String userName,
         String userAvatar,
-        byte ratingPoint) {
+        int ratingPoint) {
 }
